pyproject.toml
[project]
name = "tradepost"
version = "0.1.0"
description = "API de Maketplace Inteligente com FastAPI"
authors = [{ name = "Fillipe Berssot" }]
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "sqlalchemy[asyncio] (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic[email] (>=2.13.0,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "ruff (>=0.15.7,<0.16.0)",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
